% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.fitsujatha.R
\name{AIC.fitsujatha}
\alias{AIC.fitsujatha}
\title{Akaike Information Criterion}
\usage{
\method{AIC}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
AIC value.
}
\description{
Akaike Information Criterion
}
