% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.fitsujatha.R
\name{BIC.fitsujatha}
\alias{BIC.fitsujatha}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fitsujatha"}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A numeric value giving the BIC.
}
\description{
Returns the Bayesian Information Criterion (BIC).
}
