% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.fitsujatha.R
\name{coef.fitsujatha}
\alias{coef.fitsujatha}
\title{Coefficients of a fitted Sujatha distribution}
\usage{
\method{coef}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Estimated parameter(s).
}
\description{
Coefficients of a fitted Sujatha distribution
}
