% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.fitsujatha.R
\name{confint.fitsujatha}
\alias{confint.fitsujatha}
\title{Confidence Intervals}
\usage{
\method{confint}{fitsujatha}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{parm}{Ignored.}

\item{level}{Confidence level.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Confidence interval matrix.
}
\description{
Computes Wald confidence intervals.
}
