% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.fitsujatha.R
\name{deviance.fitsujatha}
\alias{deviance.fitsujatha}
\title{Deviance}
\usage{
\method{deviance}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Model deviance.
}
\description{
Deviance
}
