% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsujatha.R
\name{dsujatha}
\alias{dsujatha}
\title{Density Function of the Sujatha Distribution}
\usage{
dsujatha(x, theta)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{theta}{A positive parameter.}
}
\value{
A numeric vector of density values.
}
\description{
Computes the probability density function (PDF) of the
Sujatha distribution.
}
\examples{
dsujatha(1, theta = 2)
dsujatha(c(0.5, 1, 2), theta = 3)

}
