% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsujatha.R
\name{fitsujatha}
\alias{fitsujatha}
\title{Maximum Likelihood Estimation for the Sujatha Distribution}
\usage{
fitsujatha(x, start = 1, method = "L-BFGS-B")
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{start}{Initial value of theta.}

\item{method}{Optimization method (default = "L-BFGS-B").}
}
\value{
An object of class \code{"fitsujatha"} containing
\itemize{
\item Estimate
\item Standard error
\item Variance-covariance matrix
\item Hessian matrix
\item Log-likelihood
\item AIC
\item AICc
\item BIC
\item CAIC
\item HQIC
\item Deviance
\item Convergence status
\item Number of iterations
\item Original data
}
}
\description{
Fits the Sujatha distribution using Maximum Likelihood Estimation (MLE).
}
\examples{
set.seed(123)
x <- rsujatha(100, theta = 2)
fit <- fitsujatha(x)
fit

}
