% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofsujatha.R
\name{gofsujatha}
\alias{gofsujatha}
\title{Goodness-of-Fit Measures for the Sujatha Distribution}
\usage{
gofsujatha(object)
}
\arguments{
\item{object}{An object of class \code{"fitsujatha"}.}
}
\value{
A list containing
\itemize{
\item Log-likelihood
\item AIC
\item AICc
\item BIC
\item CAIC
\item HQIC
\item KS statistic
\item KS p-value
\item Anderson-Darling statistic
\item AD p-value
\item Cramer-von Mises statistic
\item CVM p-value
\item Watson statistic
\item Watson p-value
}
}
\description{
Computes several goodness-of-fit statistics for a fitted
Sujatha distribution.
}
