% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.fitsujatha.R
\name{logLik.fitsujatha}
\alias{logLik.fitsujatha}
\title{Log-likelihood}
\usage{
\method{logLik}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Object of class "logLik".
}
\description{
Log-likelihood
}
