% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitsujatha.R
\name{plot.fitsujatha}
\alias{plot.fitsujatha}
\title{Diagnostic Plots for Sujatha Distribution}
\usage{
\method{plot}{fitsujatha}(x, which = 1, ...)
}
\arguments{
\item{x}{An object of class \code{"fitsujatha"}.}

\item{which}{Integer specifying the plot to draw.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A diagnostic plot.
}
\description{
Produces diagnostic plots for an object of class
\code{"fitsujatha"}.
}
