% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fitsujatha.R
\name{print.fitsujatha}
\alias{print.fitsujatha}
\title{Print a fitted Sujatha distribution}
\usage{
\method{print}{fitsujatha}(x, ...)
}
\arguments{
\item{x}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Prints the fitted model summary.
}
\description{
Print a fitted Sujatha distribution
}
