% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psujatha.R
\name{psujatha}
\alias{psujatha}
\title{Cumulative Distribution Function of the Sujatha Distribution}
\usage{
psujatha(q, theta)
}
\arguments{
\item{q}{Numeric vector of quantiles.}

\item{theta}{Positive parameter.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Computes the cumulative distribution function (CDF).
}
\examples{
psujatha(1, theta = 2)

}
