% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.fitsujatha.R
\name{vcov.fitsujatha}
\alias{vcov.fitsujatha}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{fitsujatha}(object, ...)
}
\arguments{
\item{object}{An object of class "fitsujatha".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Variance-covariance matrix.
}
\description{
Variance-Covariance Matrix
}
